// src/components/compositions/TechnologyTag/types.ts
export type TechnologyName =
  | 'agile-methodologies'
  | 'amplitude'
  | 'apollo-client'
  | 'android-sdk'
  | 'aws'
  | 'aws-cognito'
  | 'ci-cd'
  | 'csharp'
  | 'css'
  | 'delphi'
  | 'docker'
  | 'electron'
  | 'express'
  | 'exoplayer'
  | 'firebase'
  | 'frontegg'
  | 'github-actions'
  | 'graphql'
  | 'html'
  | 'html-css'
  | 'java'
  | 'javascript'
  | 'jest'
  | 'jquery'
  | 'jspreport'
  | 'kony'
  | 'mobx'
  | 'monogame'
  | 'mysql'
  | 'nestjs'
  | 'next-js'
  | 'node-js'
  | 'opengl'
  | 'php'
  | 'postgresql'
  | 'prisma'
  | 'react'
  | 'react-hooks'
  | 'react-native'
  | 'realm'
  | 'redis'
  | 'redux'
  | 'sentry'
  | 'software-architecture'
  | 'sqlite'
  | 'sql'
  | 'sql-server'
  | 'storybook'
  | 'styled-components'
  | 'surveyjs'
  | 'swagger'
  | 'typeorm'
  | 'typescript'
  | 'unity'

export type TechnologyTagBackground = 'filled' | 'transparent'

export type TechnologyTagProps = {
  bg?: TechnologyTagBackground
  border?: 'visible' | 'none'
  showIcon?: boolean
  technology: TechnologyName
}